+++ paper/paper.bib
@@ -50,7 +50,8 @@ year = {2000}
 	publisher = {CRC press},
 	author = {Posthuma, L and {Suter II}, GW and Traas, TP},
 	year = {2001},
-	url = {https://www.crcpress.com/Species-Sensitivity-Distributions-in-Ecotoxicology/Posthuma-II-Traas/p/book/9781566705783}
+	url = {https://www.crcpress.com/Species-Sensitivity-Distributions-in-Ecotoxicology/Posthuma-II-Traas/p/book/9781566705783},
+	doi = {10.1002/etc.4373}
 }
 
 @Article{fitdistrplus,
@@ -165,13 +166,6 @@ year = {2000}
 	pages = {293--308}
 }
 
-@article{chapman_2007,
-	title = {Methods of uncertainty analysis},
-	journal={In: A H (ed) EUFRAM Concerted Action to Develop a European Framework for Probabilistic Risk Assessment of the Environmental Impacts of Pesticides, Vol 2, Detailed Reports on Role, Methods, Reporting and Validation},
-	author = {Chapman, PF RM and Hart, A and Roelofs, W and Aldenberg, T and Solomon, K and Tarazona, J LM and Byrne, P and Powley, W and Green, J and Ferson, S and Galicia, H},
-	year = {2007}
-}
-
 @techreport{fox_methodologies_2022,
 	title = {Joint investigation into {statistical} {methodologies} underpinning the {derivation} of {toxicant} {guideline} {values} in {Australia} and {New Zealand}},
 	institution = {Environmetrics Australia and Australian Institute of Marine Science},
@@ -221,9 +215,9 @@ Report prepared for the Department of Climate Change, Energy, the Environment an
 
 @techreport{USEPA2020,
    author = {{US EPA}},
-   title = {Species Sensitivity Distribution (SSD) Toolbox.},
+   title = {Species Sensitivity Distribution (SSD) Toolbox},
    year = {2020},
    publisher = { Center for Computational Toxicology and Exposure, Durham, NC},
-  url = {https://www.epa.gov/chemical-research/species-sensitivity-distribution-ssd-toolbox}
+  url = {https://www.epa.gov/chemical-research/species-sensitivity-distribution-ssd-toolbox},
+  doi = {10.23645/epacomptox.11971392.v2}
 }
-
