### 2024-10-18 53746e0
set pkgdown development mode auto
