+++ _pkgdown.yml
@@ -125,3 +125,6 @@ articles:
   - articles/confidence-intervals
   - customising-plots
   - additional-technical-details
+
+development:
+  mode: auto
