### 2024-10-18 9d41ea6
remove https://doi.org from doi
