+++ paper/paper.bib
@@ -6,7 +6,7 @@ volume = {19},
 number = {2},
 pages = {508-515},
 keywords = {Risk assessment, Species sensitivity, Bootstrap, Statistics, Sample size},
-doi = {https://doi.org/10.1002/etc.5620190233},
+doi = {10.1002/etc.5620190233},
 url = {https://setac.onlinelibrary.wiley.com/doi/abs/10.1002/etc.5620190233},
 eprint = {https://setac.onlinelibrary.wiley.com/doi/pdf/10.1002/etc.5620190233},
 abstract = {Abstract Species-sensitivity distribution methods assemble single-species toxicity data to predict hazardous concentrations (HCps) affecting a certain percentage (p) of species in a community. The fit of the lognormal model and required number of individual species values were evaluated with 30 published data sets. The increasingly common assumption that a lognormal model best fits these data was not supported. Fifteen data sets failed a formal test of conformity to a lognormal distribution; other distributions often provided better fit to the data than the lognormal distribution. An alternate bootstrap method provided accurate estimates of HCp without the assumption of a specific distribution. Approximate sample sizes producing HC5 estimates with minimal variance ranged from 15 to 55, and had a median of 30 species-sensitivity values. These sample sizes are higher than those suggested in recent regulatory documents. A bootstrap method is recommended that predicts with 95\% confidence the concentration affecting 5\% or fewer species.},
