### 2024-10-18 b1b0279
skip on cran for weibull failing on M1mac likely due to much older version of the OS and toolchain
