### 2024-10-18 0b97a68
- Set ssd_xxmulti(lnorm.weight = 0)` (instead of 1) to avoid incorrect values
which can occur with `do.call("ssd_xxmulti", c(n = n, estimates(fits, all_estimates = FALSE))`
