+++ tests/testthat/test-multi.R
@@ -16,21 +16,30 @@
 #    limitations under the License.
 
 test_that("multi", {
-  test_dist("multi", multi = TRUE)
-  expect_equal(ssd_pmulti(1), 0.5)
-  expect_equal(ssd_qmulti(0.75), 1.96303108415826)
+  expect_equal(ssd_pmulti(1, lnorm.weight = 0.5), 0.5)
+  expect_equal(ssd_qmulti(0.75, lnorm.weight = 2), 1.96303108415826)
   set.seed(42)
-  expect_equal(ssd_rmulti(2), c(3.93912428813385, 4.62130564767823))
+  expect_equal(ssd_rmulti(2, lnorm.weight = 1), c(3.93912428813385, 4.62130564767823))
+  
+  set.seed(42)
+  expect_equal(ssd_rmulti(1, gamma.weight = 0.5, lnorm.weight = 0.5), 3.13234340623737)
 
-  expect_equal(ssd_qmulti(ssd_pmulti(c(0, 0.1, 0.5, 0.9, 0.99))),
+  set.seed(42)
+  expect_equal(ssd_rmulti(1, gamma.weight = 1, lnorm.weight = 1), 3.13234340623737)
+  
+  expect_equal(ssd_qmulti(ssd_pmulti(c(0, 0.1, 0.5, 0.9, 0.99), lnorm.weight = 1), lnorm.weight = 1),
     c(0, 0.1, 0.5, 0.9, 0.99),
     tolerance = 1e-5
   )
 
-  expect_equal(ssd_pmulti(ssd_qmulti(c(0, 0.1, 0.5, 0.9, 0.99))),
+  expect_equal(ssd_pmulti(ssd_qmulti(c(0, 0.1, 0.5, 0.9, 0.99), lnorm.weight = 1), lnorm.weight = 1),
     c(0, 0.1, 0.5, 0.9, 0.99),
     tolerance = 1e-6
   )
+  expect_error(ssd_pmulti(0.5), "^At least one distribution must have a positive weight\\.$")
+  expect_error(ssd_qmulti(0.75), "^At least one distribution must have a positive weight\\.$")
+  expect_error(ssd_rmulti(1), "^At least one distribution must have a positive weight\\.$")
+  test_dist("multi", multi = TRUE)
 })
 
 test_that("ssd_pmulti", {
