+++ man/ssd_r.Rd
@@ -78,7 +78,7 @@ ssd_rmulti(
   llogis_llogis.locationlog2 = 1,
   llogis_llogis.scalelog2 = 1,
   llogis_llogis.pmix = 0.5,
-  lnorm.weight = 1,
+  lnorm.weight = 0,
   lnorm.meanlog = 0,
   lnorm.sdlog = 1,
   lnorm_lnorm.weight = 0,
@@ -269,7 +269,7 @@ hist(ssd_rlnorm(10000), breaks = 1000)
 
 # multi
 set.seed(50)
-hist(ssd_rmulti(1000), breaks = 100)
+hist(ssd_rmulti(1000, gamma.weight = 0.5, lnorm.weight = 0.5), breaks = 100)
 
 fits <- ssd_fit_dists(ssddata::ccme_boron)
 do.call("ssd_rmulti", c(n = 10, estimates(fits)))

[…]
