+++ NEWS.md
@@ -1,5 +1,10 @@
 <!-- NEWS.md is maintained by https://fledge.cynkra.com, contributors should not edit this file -->
 
+# ssdtools 2.0.0.9000
+
+- Set `ssd_xxmulti(lnorm.weight = 0)` (instead of 1) to avoid incorrect values.
+
+
 # ssdtools 2.0.0
 
 `ssdtools` v2.0.0, which now includes David Fox and Rebecca Fisher as co-authors, is the second major release of `ssdtools`.
