### 2024-10-18 012bb82
all estimates not unnecessary in do.calls for ssd_xx
