### 2024-10-18 fe34ece
- Add `ssd_xxmulti_fitdists()` functions to accept object of class fitdists.
