+++ R/multi.R
@@ -200,9 +200,6 @@ ssd_qmulti <- function(
 #' # multi
 #' set.seed(50)
 #' hist(ssd_rmulti(1000, gamma.weight = 0.5, lnorm.weight = 0.5), breaks = 100)
-#'
-#' fits <- ssd_fit_dists(ssddata::ccme_boron)
-#' do.call("ssd_rmulti", c(n = 10, estimates(fits)))
 ssd_rmulti <- function(
     n,
     burrIII3.weight = 0,
@@ -296,7 +293,15 @@ ssd_emulti <- function() {
   as.list(unlist(emulti))
 }
 
-.ssd_pmulti_fitdists <- function(q, fitdists, lower.tail = TRUE, log.p = FALSE) {
+#' @describeIn ssd_p Cumulative Distribution Function for Multiple Distributions
+#' @export
+#' @examples
+#'
+#' # multi fitdists
+#' fit <- ssd_fit_dists(ssddata::ccme_boron)
+#' ssd_pmulti_fitdists(1, fit)
+ssd_pmulti_fitdists <- function(q, fitdists, lower.tail = TRUE, log.p = FALSE) {
+  chk_s3_class(fitdists, "fitdists")
   args <- estimates(fitdists)
   args$q <- q
   args$lower.tail <- lower.tail
@@ -304,7 +309,15 @@ ssd_emulti <- function() {
   do.call("ssd_pmulti", args)
 }
 
-.ssd_qmulti_fitdists <- function(p, fitdists, lower.tail = TRUE, log.p = FALSE) {
+#' @describeIn ssd_q Quantile Function for Multiple Distributions
+#' @export
+#' @examples
+#'
+#' # multi fitdists
+#' fit <- ssd_fit_dists(ssddata::ccme_boron)
+#' ssd_qmulti_fitdists(0.5, fit)
+ssd_qmulti_fitdists <- function(p, fitdists, lower.tail = TRUE, log.p = FALSE) {
+  chk_s3_class(fitdists, "fitdists")
   args <- estimates(fitdists)
   args$p <- p
   args$lower.tail <- lower.tail
@@ -312,7 +325,15 @@ ssd_emulti <- function() {
   do.call("ssd_qmulti", args)
 }
 
-.ssd_rmulti_fitdists <- function(n, fitdists, chk = TRUE) {
+#' @describeIn ssd_r Random Generation for Multiple Distributions
+#' @export
+#' @examples
+#'
+#' # multi fitdists
+#' fit <- ssd_fit_dists(ssddata::ccme_boron)
+#' ssd_rmulti_fitdists(2, fit)
+ssd_rmulti_fitdists <- function(n, fitdists, chk = TRUE) {
+  chk_s3_class(fitdists, "fitdists")
   args <- estimates(fitdists)
   args$n <- n
   args$chk <- chk
