+++ tests/testthat/test-multi.R
@@ -44,19 +44,19 @@ test_that("multi", {
 
 test_that("ssd_pmulti", {
   fit <- ssd_fit_dists(data = ssddata::ccme_boron)
-  expect_identical(.ssd_pmulti_fitdists(numeric(0), fit), numeric(0))
-  expect_identical(.ssd_pmulti_fitdists(NA_real_, fit), NA_real_)
-  expect_identical(.ssd_pmulti_fitdists(-Inf, fit), 0)
-  expect_equal(.ssd_pmulti_fitdists(Inf, fit), 1)
-  expect_equal(.ssd_pmulti_fitdists(0, fit), 0)
+  expect_identical(ssd_pmulti_fitdists(numeric(0), fit), numeric(0))
+  expect_identical(ssd_pmulti_fitdists(NA_real_, fit), NA_real_)
+  expect_identical(ssd_pmulti_fitdists(-Inf, fit), 0)
+  expect_equal(ssd_pmulti_fitdists(Inf, fit), 1)
+  expect_equal(ssd_pmulti_fitdists(0, fit), 0)
   pone <- 0.0389879276872944
-  expect_equal(.ssd_pmulti_fitdists(1, fit), pone, tolerance = 1e-5)
-  expect_equal(.ssd_pmulti_fitdists(10000, fit), 0.999954703139271, tolerance = 1e-6)
-  expect_equal(.ssd_pmulti_fitdists(c(1, 2), fit), c(pone, 0.0830184001863268), tolerance = 1e-5)
-  expect_equal(.ssd_pmulti_fitdists(c(1, NA), fit), c(pone, NA), tolerance = 1e-5)
-  expect_equal(.ssd_pmulti_fitdists(1, fit, lower.tail = FALSE), 1 - pone, tolerance = 1e-6)
-  expect_equal(.ssd_pmulti_fitdists(1, fit, log.p = TRUE), log(pone), tolerance = 1e-6)
-  expect_equal(.ssd_pmulti_fitdists(1, fit, lower.tail = FALSE, log.p = TRUE), log(1 - pone), tolerance = 1e-5)
+  expect_equal(ssd_pmulti_fitdists(1, fit), pone, tolerance = 1e-5)
+  expect_equal(ssd_pmulti_fitdists(10000, fit), 0.999954703139271, tolerance = 1e-6)
+  expect_equal(ssd_pmulti_fitdists(c(1, 2), fit), c(pone, 0.0830184001863268), tolerance = 1e-5)
+  expect_equal(ssd_pmulti_fitdists(c(1, NA), fit), c(pone, NA), tolerance = 1e-5)
+  expect_equal(ssd_pmulti_fitdists(1, fit, lower.tail = FALSE), 1 - pone, tolerance = 1e-6)
+  expect_equal(ssd_pmulti_fitdists(1, fit, log.p = TRUE), log(pone), tolerance = 1e-6)
+  expect_equal(ssd_pmulti_fitdists(1, fit, lower.tail = FALSE, log.p = TRUE), log(1 - pone), tolerance = 1e-5)
 })
 
 test_that("ssd_pmulti weights", {
@@ -80,19 +80,19 @@ test_that("ssd_pmulti weights", {
 
 test_that("ssd_qmulti", {
   fit <- ssd_fit_dists(data = ssddata::ccme_boron)
-  expect_identical(.ssd_qmulti_fitdists(numeric(0), fit), numeric(0))
-  expect_identical(.ssd_qmulti_fitdists(NA_real_, fit), NA_real_)
-  expect_identical(.ssd_qmulti_fitdists(-1, fit), NaN)
-  expect_identical(.ssd_qmulti_fitdists(-Inf, fit), NaN)
[…]
