+++ dev/pkgdown.yml
@@ -1,0 +1,11 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  additional-technical-details: additional-technical-details.html
+  articles/confidence-intervals: confidence-intervals.html
+  customising-plots: customising-plots.html
+  distributions: distributions.html
+  model-averaging: model-averaging.html
+  ssdtools: ssdtools.html
+last_built: 2024-10-18T21:00Z
