+++ R/exposure.R
@@ -4,7 +4,7 @@
 # Energy, the Environment and Water
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compl %>% iance with the License.
+#    you may not use this file except in compliance with the License.
 #    You may obtain a copy of the License at
 #
 #       https://www.apache.org/licenses/LICENSE-2.0
