+++ NEWS.md
@@ -2,23 +2,8 @@
 
 # ssdtools 2.1.0
 
-- Merge pull request #385 from poissonconsulting/dev.
-
-  fix cran and bug if using multi functions with lnorm not included
-
-- Merge pull request #384 from bcgov/cran.
-
-  cran v2.0.0
-
-
-# ssdtools 2.0.0.9001
-
-- Add `ssd_xxmulti_fitdists()` functions to accept object of class fitdists.
-
-
-# ssdtools 2.0.0.9000
-
-- Set `ssd_xxmulti(lnorm.weight = 0)` (instead of 1) to avoid incorrect values with `do.call("ssd_xxmulti", c(..., estimates(fits))` if `fits` does not include the log-normal distribution.
+- Added `ssd_xxmulti_fitdists()` functions to accept object of class `fitdists`.
+- Set `ssd_xxmulti(..., lnorm.weight = 0)` (instead of 1) to avoid incorrect values with `do.call("ssd_xxmulti", c(..., estimates(fits))` if `fits` does not include the log-normal distribution.
 
 
 # ssdtools 2.0.0
