+++ cran-comments.md
@@ -12,4 +12,4 @@ The large size of these sub-directories is necessary.
 
 ## CRAN Package Check Results for Package
 
-Fixed noRemap issue.
+Fixed M1mac issue.
