### 2024-10-20 #7351 @Kevin-Mattheus-Moerman
Note to editors, a prior paper for this project exists: https://joss.theoj.org/papers/10.21105/joss.01082
