### 2024-10-20 #7351 @Kevin-Mattheus-Moerman
@joethorley you noted that this new submission (wrt the previous paper) now also features changes like:

> The first update (v1) included the addition of four new distributions and a switch to the R package TMB. The second major release (v2) includes critical updates to ensure that the HC and HP estimates satisfy the inversion principle as well as bootstrap methods to obtain confidence intervals (CIs) with more appropriate coverage.

If you could share a bit more detail on the major changes since the previous paper, e.g. a list here, that would be helpful for the handling editor. Thanks.
