+++ README.Rmd
@@ -58,7 +58,7 @@ pak::pak("bcgov/ssdtools")
 
 #### Website
 
-The website for the development version is at <https://bcgov.github.io/ssdtools/dev>.
+The website for the development version is at <https://bcgov.github.io/ssdtools/dev/>.
 
 ## Introduction
 
