+++ dev/pkgdown.yml
@@ -8,4 +8,4 @@ articles:
   distributions: distributions.html
   model-averaging: model-averaging.html
   ssdtools: ssdtools.html
-last_built: 2024-10-18T21:00Z
+last_built: 2024-10-21T21:28Z
