+++ pkgdown.yml
@@ -8,4 +8,4 @@ articles:
   distributions: distributions.html
   model-averaging: model-averaging.html
   ssdtools: ssdtools.html
-last_built: 2024-10-10T15:19Z
+last_built: 2024-10-22T14:53Z
