### 2024-10-22 #7351 @joethorley
@Kevin-Mattheus-Moerman The following is a  relatively short summary of the major changes since the previous paper (v0.0.3)

## ssdtools 2.0.0

The following arguments were added to `ssd_hc()` and `ssd_hp()`

  - `multi_est = TRUE` to calculate model averaged estimates treating the distributions as constituting a single mixture distribution (previously it was effectively `FALSE`).
  - `method_ci = "weighted_samples"` to specify whether to use `"weighted_samples"`, `"weighted_arithmetic"`, `"multi_free"` or `"multi_fixed"` methods to generate confidence intervals (previously it was effectively `"weighted_arithmetic"`).

In addition, the following functions and arguments were added.

- `scale_fill_ssd()` for color-blind fill scale.

## ssdtools 1.0.0

An important change to the functionality of `ssd_fit_dists()` was to switch from model fitting using [`fitdistrplus`](https://github.com/lbbe-software/fitdistrplus) to [`TMB`](https://github.com/kaskr/adcomp) which has resulted in improved handling of censored data.
Although it was hoped that model fitting would be faster this is currently not the case.

As a result of an international collaboration British Columbia and Canada and Australia and New Zealand selected a set of recommended distributions for model averaging and settings when generating final guidelines.

The distributions are
```{r}
> ssd_dists_bcanz()
[1] "gamma"       "lgumbel"     "llogis"      "lnorm"       "lnorm_lnorm" "weibull"
```

The following distributions were added (or in the case of `burrIII3` readded) to the new version

  - `burrIII3` - burrIII three parameter distribution
  - `invpareto` - inverse pareto (with bias correction in scale order statistic)
  - `lnorm_lnorm` log-normal/log-normal mixture distribution
  - `llogis_llogis` log-logistic/log-logistic mixture distribution

The function `ssd_fit_burrlioz()` was added to approximate the behaviour of [Burrlioz](https://research.csiro.au/software/burrlioz/).

Added following plotting functions

  - `geom_ssdsegment()` to allow plotting of the range of a censored data points using segments.
  - `scale_colour_ssd()` (and `scale_color_ssd()`) to provide an 8 color-blind scale.

Soft-deprecated
  - `geom_ssd()` for `geom_ssdpoint()`.

## ssdtools 0.3.0

- Soft-deprecated `pareto` distribution as poor fit on SSD data.

## ssdtools 0.2.0

- Deprecated `burrIII2` distribution for `llogis` as identical.
- Replaced `burrIII2` with (identical) `llogis` distribution in default set.

## ssdtools 0.1.0

- Default distributions changed to `burrIII2`, `gamma` and `lnorm` from
`gamma`, `gompertz`, `lgumbel`, `llog`, `lnorm` and `weibull`.

Added
  - Burr Type-III Two-Parameter Distribution (`burrIII2`).
  - Added Burr Type-III Three-Parameter Distribution (`burrIII3`).
