+++ vignettes/additional-technical-details.Rmd
@@ -66,7 +66,7 @@ A total of 20 unique datasets were extracted from ssddata and used to define the
 
 The individual ssdtools fits are shown below for each of the 20 simulation datasets from ssddata, for the six recommended default distributions, as well as the model averaged CDF (black line):
 
-![](Images/fitted_dists.png)
+![](images/fitted_dists.png)
 
 This simulation process was repeated a minimum of 1,000 times for each dataset, and the results collated across all iterations. 
 For each simulated dataset the true HCx values were obtained directly from the parameter estimates of from data generating distribution.
@@ -74,7 +74,7 @@ From these, relative bias was calculated as the scaled-difference between the es
 
 Bias, confidence interval width and coverage as a function of sample size across ~1000 simulations of 20 datasets using the multi model averaging method and the weighted sample method for estimating confidence intervals via ssdtools are shown below:
 
-![](Images/ssdata_sims_collated.png)
+![](images/ssdata_sims_collated.png)
 
 The simulation results showed significant gains in terms of reduced bias from N=5 to N=6, as well as in coverage, which improved substantially between N=5 and N=6. 
 There is also a small additional gain in coverage at N=7, where the median of the simulations reaches the expected 95% but this is only the case for HC1.
@@ -88,7 +88,7 @@ Estimates of coverage and confidence interval widths were not obtained for this 
 
 Bias of sample size across ~1000 simulations of 353 datasets using the multi model averaging method via ssdtools, for HC5, 10 and 20 (0.05, 0.1, and 0.2) are shown below:
 
-![](Images/all_sims_bias.png)
+![](images/all_sims_bias.png)
 
 Note that simulation results are shown separately for those derived from each of the six distributions as the underlying source data generating distribution.
 
@@ -105,7 +105,7 @@ This was examined using the simulation study across the larger combined ssddata 
 Below is a plot of the mean AICc weights as a function of sample size (N) obtained for data simulated using the best fit distribution to 353 datasets. 
 Results are shown separately for the six different source distributions, with the upper plot (A) showing the AICc weight of the source (data generating) distributions, when fit using the default set of six distributions using ssdtools; and the lower plot (B) showing the AICc weight of the lognormal-lognormal mixture distribution for each of the source (data generating) distributions.
 
-![](Images/weights_collated.png)
+![](images/weights_collated.png)
 
 We found that the AICc weights for the five unimodal distributions were relatively similar (~0.2) for very low N. 
 This is because for small N it is difficult to discern differences between the distributions in the candidate list.
