### 2024-10-24 bcgov/ssdtools#387 @beckyfisher
Some equations not rendering properly on github pages
Some equations not rendering properly on github website
![image](https://github.com/user-attachments/assets/0dff94eb-3269-452c-8b6b-5541d7b8749e)
