### 2024-10-24 cab39f1
Correction to bimodal distributions y axis labels
