### 2024-10-24 03a636a
tweaks to equation code to address #387
