+++ vignettes/additional-technical-details.Rmd
@@ -130,20 +130,25 @@ Based on these results, our recommendation is that only a single minimum sample 
 
 ### Burr III distribution
 
-The probability density function, ${f_X}(x;b,c,k)$ and cumulative distribution function, ${F_X}(x;b,c,k)$ for the Burr III distribution (also known as the *Dagum* distribution) as used in `ssdtools` are:
+The probability density function, ${f_X}(x;b,c,k)$ and cumulative distribution function, ${F_X}(x;b,c,k)$ for the Burr III distribution (also known as the *Dagum* distribution) as used in `ssdtools` are
 
 <div style="text-align: center;"font-size: 1.2em; color:#3664EE;" class="box";>
-<b><u>Burr III distribution</u></b>\[\begin{array}{*{20}{c}}
+<b><u>Burr III distribution</u></b>
+
+\[\begin{array}{*{20}{c}}
 {{f_X}(x;b,c,k) = \frac{{b\,k\,c}}{{{x^2}}}\frac{{{{\left( {\frac{b}{x}} \right)}^{c - 1}}}}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^{k + 1}}}}{\rm{ }}}&{b,c,k,x > 0}
 \end{array}\]<br>\[\begin{array}{*{20}{c}}
 {{F_X}(x;b,c,k) = \frac{1}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^k}}}{\rm{ }}}&{b,c,k,x > 0}
 \end{array}\]
+
 </div>
 
 ### Inverse Pareto distribution
 
-<br>Let $X \sim Burr(b,c,k)$ have the *pdf* given in the box above. 
-It is well known that the distribution of $Y = \frac{1}{X}$ is the *inverse Burr* distribution (also known as the *SinghMaddala* distribution) for which:\[\begin{array}{*{20}{c}}
+Let $X \sim Burr(b,c,k)$ have the *pdf* given in the box above. 
+It is well known that the distribution of $Y = \frac{1}{X}$ is the *inverse Burr* distribution (also known as the *SinghMaddala* distribution) for which 
+
+\[\begin{array}{*{20}{c}}
 {{f_Y}(y;b,c,k) = \frac{{c{\kern 1pt} {\kern 1pt} k{{\left( {\frac{y}{b}} \right)}^c}}}{{y{\kern 1pt} {{\left[ {1 + {{\left( {\frac{y}{b}} \right)}^c}} \right]}^{k + 1}}}}}&{b,c,k,y > 0}
 \end{array}\]
 
@@ -172,33 +177,38 @@ Therefore, \[\begin{array}{*{20}{c}}
 {\mathop {\mathop {\lim }\limits_{(c,k) \to (\infty ,0)} }\limits_{ck = \lambda } \left\{ {{F_Y}(y;b,c,k)} \right\} = 1 - {{\left( {\frac{b}{y}} \right)}^\lambda }}&{y \ge b}
 \end{array}\]
 
-which we recognise as the (American) Pareto distribution. So, if the limiting distribution of $Y = \frac{1}{X}$ is a Pareto distribution, then the limiting distribution of $X = \frac{1}{Y}$ is the (American) *inverse Pareto* distribution:
[…]
