### 2024-11-13 #7351 @flor14
Hello,

Regarding the review process: Is there any difference in the reviewers' checklist for a second version of a software package that has already been published? I couldn’t find any mention of this in the reviewers' guide.
