### 2024-11-18 #7492 @editorialbot
[REVIEW]: ssdtools v2: An R package to fit Species Sensitivity Distributions
**Submitting author:** @joethorley (<a href="http://orcid.org/0000-0002-7683-4592">Joseph Thorley</a>)
**Repository:** https://github.com/bcgov/ssdtools
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.2.0
