### 2024-11-18 #7492 @fabian-s
@joethorley while we wait for the reviews, could you check the "missing DOIs" the bot found in the comment above and add them to your refs if they are correct?
