### 2024-11-20 #7492 @joethorley
> @joethorley while we wait for the reviews, could you check the "missing DOIs" the bot found in the comment above and add them to your refs if they are correct?

@fabian-s I've added the two missing DOIs.
