### 2024-11-21 #7492 @flor14
@fabian-s
My review is complete. I believe the package is truly valuable, and these updates position it as the most up-to-date and comprehensive software for fitting and calculating relevant concentrations from an SSD. I also appreciate the work put into the documentation, which is easy to follow and well-organized.

One aspect I’ve always appreciated about JOSS papers is that they provide insights into aspects of the software not covered in other sections of the documentation. Since this is the second article for _ssdtools_, I found myself considering what additional information could be included, especially given that there is already an existing article for this software. For a manuscript's second version, it could be valuable to include examples addressing changes in functions that result in breaking changes. Also, requesting a more detailed explanation of the reasons and the impacts of changes in dependencies could be informative in some cases. However, I understand that this may go beyond the scope of this review as the process is now, so I focused on the points outlined in the checklist. Probably, you have already considered this when deciding to allow the publication of new versions and have taken these factors into account. However, I wanted to mention it since it influenced the comments I made in my review.

@joethorley Only two points of the review have not been marked as resolved. I added notes on a few other points, but there are suggestions or comments that I don't consider to affect the quality of the software/manuscript. I believe these are "stylistic" considerations that are up to you, so I marked them as solved.
