### 2024-11-21 #7492 @fabian-s
@flor14 thank you so much for your timely, constructive and detailed review -- this is great! I second your recommendations, with one exception. i.e. "It might be helpful to organize the Technical Details section to indicate which changes were introduced in the first major version update and which were added in the second one."  -- this level of historical detail seems less relevant to me, since the new paper is supposed to  document the *current* state of the software.
