### 2024-11-22 9f95b9f
fix bug in scale_colour_ssd() example
