### 2024-11-22 #7492 @flor14
Thank you, @fabian-s, for your kind words and for reading my comments. You're right. Sometimes I like to include more detailed ways of organizing text to make the information more accessible to readers. I acknowledge that perhaps it wasn’t the best example. Either way, it’s one of the points the review marked as complete. Thank you for your feedback, and have a great weekend!
