### 2024-11-22 #7492 @nanhung
I just completed my review of the package. It is well-developed, actively maintained, and widely adopted in recent publications, which speaks to its reliability and usefulness. One potential enhancement would be adding a comprehensive set of examples to simplify variable manipulation for users. Detailed comments are listed below:

- Given the availability of packages like 'viridis' designed for color-blind accessibility, any specific purpose of `scale_color_ssd()`? Additionally, since it only offers eight colors, what would you recommend for datasets with more than eight groups?
- There are currently no examples illustrating `scale_fill_ssd()` in SSD analysis. What is the application of this function?
- The [description](https://bcgov.github.io/ssdtools/reference/ssd_censor_data.html) only states "Censor Data". I suggest adding further explanation of its purpose and how it fits into typical workflows?
- I guess `multi_est` is related to `ci_method`. The method for integrating `multi_est` and `ci_method` is unclear. Additional guidance on their use would be helpful.
- In the Model Fitting section, "The move to TMB allowed more control over model specification" is vague. More details are needed to explain what option that can be used in model specification.
- In the section of Multiple Distribution Function. It might need more explanation about how these functions implement in the "multi" method.
- The function ssd_xxmulti_fitdists() is referenced in the [ChangeLog](https://bcgov.github.io/ssdtools/news/index.html#ssdtools-210) but does not appear in the package.
- Some wanrnings were found when using `pkgdown` building package site.

[
![Screenshot from 2024-11-22 08-55-43](https://github.com/user-attachments/assets/7deceaa1-1acc-4a15-8383-f336b530b95b)
](url)

- All pictures not display in [Additional Technical Details](https://bcgov.github.io/ssdtools/articles/additional-technical-details.html)

![Screenshot from 2024-11-22 14-07-50](https://github.com/user-attachments/assets/6ad3b53c-18c0-4136-91bc-abe036d10149)

I will open PR if I find any bug that can be simply fixed. Overall, nice package with a lot of refinements!
