### 2024-11-25 #7492 @fabian-s
@nanhung thank you as well for a fantastic (and quick!) job with this review. reviewers like you and @flor14 make JOSS possible and this kind of peer review worthwhile!
