### 2024-11-25 #7492 @fabian-s
@joethorley both reviewers make a number of suggestions to improve the paper/the software that seem very sensible to me.
please adress them point by point, either here in the thread or by opening (& closing, hopefully)  relevant issues in your repo (e.g. https://github.com/bcgov/ssdtools/issues/182 re. the alt text in the vignettes). Please let us know about your timeframe to adress them.
