+++ doc/model-averaging.R
@@ -1,0 +1,173 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 6,
+  fig.height = 4
+)
+library(tinytex)
+
+## ----include=FALSE------------------------------------------------------------
+# this just loads the package and suppresses the load package message
+library(ssdtools)
+require(ggplot2)
+
+## ----echo=FALSE,warning=FALSE, message=FALSE,class.output="scroll-100"--------
+samp <- c(1.73, 0.57, 0.33, 0.28, 0.3, 0.29, 2.15, 0.8, 0.76, 0.54, 0.42, 0.83, 0.21, 0.18, 0.59)
+print(samp)
+# knitr::kable(samp,caption="Some toxicity data (concentrations)")
+
+## ----echo=FALSE,fig.cap="Emprirical cdf (black); Model 1(green); and Model 2 (blue)", fig.width=7,fig.height=4.5----
+samp <- c(1.73, 0.57, 0.33, 0.28, 0.3, 0.29, 2.15, 0.8, 0.76, 0.54, 0.42, 0.83, 0.21, 0.18, 0.59)
+samp <- sort(samp)
+plot(ecdf(samp), main = "Empirical and fitted SSDs", xlab = "Concentration", ylab = "Probability")
+xx <- seq(0.01, 3, by = 0.01)
+lines(xx, plnorm(xx, meanlog = mean(log(samp[1:10])), sd = sd(log(samp[1:10]))), col = "#77d408")
+lines(xx, plnorm(xx, meanlog = mean(log(samp[5:15])), sd = sd(log(samp[5:15]))), col = "#08afd4")
+# lines(xx,(0.4419*plnorm(xx,meanlog=mean(log(samp[5:15])),sd=sd(log(samp[5:15])))+
+#          0.5581*plnorm(xx,meanlog=mean(log(samp[1:10])),sd=sd(log(samp[1:10]))))  ,col="#d40830")
+
+## ----echo=FALSE,fig.cap="Empirical cdf (black); Model 1(green); Model 2 (blue); and averaged Model (red)",fig.width=7,fig.height=4.5----
+samp <- c(1.73, 0.57, 0.33, 0.28, 0.3, 0.29, 2.15, 0.8, 0.76, 0.54, 0.42, 0.83, 0.21, 0.18, 0.59)
+samp <- sort(samp)
+plot(ecdf(samp), main = "Empirical and fitted SSDs", xlab = "Concentration", ylab = "Probability")
+xx <- seq(0.01, 3, by = 0.01)
+lines(xx, plnorm(xx, meanlog = mean(log(samp[1:10])), sd = sd(log(samp[1:10]))), col = "#77d408")
+lines(xx, plnorm(xx, meanlog = mean(log(samp[5:15])), sd = sd(log(samp[5:15]))), col = "#08afd4")
+lines(xx, (0.4419 * plnorm(xx, meanlog = mean(log(samp[5:15])), sd = sd(log(samp[5:15]))) +
+  0.5581 * plnorm(xx, meanlog = mean(log(samp[1:10])), sd = sd(log(samp[1:10])))), col = "#d40830")
+
+## ----echo=TRUE----------------------------------------------------------------
+# Model 1 HC20
+cat("Model 1 HC20 =", qlnorm(0.2, -1.067, 0.414))
+
+# Model 2 HC20
+cat("Model 2 HC20 =", qlnorm(0.2, -0.387, 0.617))
+
+## ----echo=FALSE,fig.width=7,fig.height=5--------------------------------------
[…]
