+++ doc/distributions.R
@@ -1,0 +1,315 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 6,
+  fig.height = 4
+)
+
+## ----echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample Burr probability density (A) and cumulative probability (B) functions."----
+par(mfrow = c(1, 2))
+
+f <- function(x, b, c, k) {
+  z1 <- (b / x)^(c - 1)
+  z2 <- (b / x)^c
+  y <- (b * c * k / x^2) * z1 / (1 + z2)^(k + 1)
+  return(y)
+}
+
+conc <- seq(0, 10, by = 0.005)
+plot(conc, f(conc, 1, 3, 5), type = "l", ylab = "Probability density", xlab = "Concentration", col = "#FF5733")
+lines(conc, f(conc, 1, 1, 2), col = "#F2A61C")
+lines(conc, f(conc, 1, 2, 2), col = "#1CADF2")
+lines(conc, f(conc, 1, 2, 5), col = "#1F1CF2")
+legend("topleft", "(A)", bty = "n")
+
+F <- function(x, b, c, k) {
+  z2 <- (b / x)^c
+  y <- 1 / (1 + z2)^k
+  return(y)
+}
+conc <- seq(0, 10, by = 0.005)
+plot(conc, F(conc, 1, 3, 5), type = "l", ylab = "Cumulative probability", xlab = "Concentration", col = "#FF5733")
+lines(conc, F(conc, 1, 1, 2), col = "#F2A61C")
+lines(conc, F(conc, 1, 2, 2), col = "#1CADF2")
+lines(conc, F(conc, 1, 2, 5), col = "#1F1CF2")
+legend("topleft", "(B)", bty = "n")
+
+## ----echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample lognormal lognormal mixture probability density (A) and cumulative probability (B) functions."----
+par(mfrow = c(1, 2))
+
+f <- function(x, m1, s1, m2, s2, p) {
+  y <- p * dlnorm(x, m1, s1) + (1 - p) * dlnorm(x, m2, s2)
+  return(y)
+}
+
+conc <- seq(0, 5, by = 0.0025)
+m1 <- 1
+s1 <- .2
+m2 <- 1.8
+s2 <- 1.5
+p <- 0.25
+plot(conc, f(conc, m1, s1, m2, s2, p), type = "l", ylab = "Probability density", xlab = "Concentration", col = "#FF5733", ylim = c(0, 1))
+lines(conc, f(conc, 0.09, 0.5, 1, 0.08, 0.9), col = "#F2A61C")
+lines(conc, f(conc, 0.5, 0.5, 1, 0.08, 0.9), col = "#1CADF2")
+lines(conc, f(conc, 0.7, 1.5, 0.5, 0.1, .7), col = "#1F1CF2")
+legend("topleft", "(A)", bty = "n")
+
+F <- function(x, m1, s1, m2, s2, p) {
+  y <- p * plnorm(x, m1, s1) + (1 - p) * plnorm(x, m2, s2)
+  return(y)
+}
+
+conc <- seq(0, 10, by = 0.0025)
+m1 <- 1
+s1 <- .2
+m2 <- 1.8
+s2 <- 1.5
+p <- 0.25
+plot(conc, F(conc, m1, s1, m2, s2, p), type = "l", ylab = "Cumulative probability", xlab = "Concentration", col = "#FF5733", ylim = c(0, 1))
[…]
