+++ doc/additional-technical-details.R
@@ -1,0 +1,12 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 6,
+  fig.height = 4,
+  echo = TRUE
+)
+
+## ----results = "asis", echo = FALSE-------------------------------------------
+cat(ssdtools::ssd_licensing_md())
+
