+++ doc/small-sample-bias.R
@@ -1,0 +1,1 @@
+### This is an R script tangled from 'small-sample-bias.pdf.asis'

[…]
