### 2024-12-02 2d387af
distributions.Rmd latex boxes
