+++ vignettes/distributions.Rmd
@@ -76,34 +76,18 @@ New Zealand Guidelines 2018).
 
 The probability density function, ${f_X}(x;b,c,k)$ and cumulative distribution function, ${F_X}(x;b,c,k)$ for the Burr III distribution are:
 
-<!-- <div style="text-align: center;"font-size: 1.2em; color:#3664EE;" class="box";> -->
-<!-- <b><u>Burr III distribution</u></b>\[\begin{array}{*{20}{c}} -->
-<!-- {{f_X}(x;b,c,k) = \frac{{b\,k\,c}}{{{x^2}}}\frac{{{{\left( {\frac{b}{x}} \right)}^{c - 1}}}}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^{k + 1}}}}{\rm{ }}}&{b,c,k,x > 0} -->
-<!-- \end{array}\]<br>\[\begin{array}{*{20}{c}} -->
-<!-- {{F_X}(x;b,c,k) = \frac{1}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^k}}}{\rm{ }}}&{b,c,k,x > 0} -->
-<!-- \end{array}\] -->
-<!-- </div> -->
-
-<div style="text-align: center; font-size: 1.2em; color: #3664EE; border: 1px solid #3664EE; padding: 10px; border-radius: 5px;">
+<div style="text-align: center; font-size: 1.2em; color: black; border: 1px solid black; padding: 10px; border-radius: 5px;">
   <b><u>Burr III Distribution</u></b>
   
-  $$f_X(x;b,c,k) = \frac{{b\,k\,c}}{{x^2}} \frac{{\left( \frac{b}{x} \right)}^{c - 1}}{{\left[ 1 + \left( \frac{b}{x} \right)^c \right]}^{k + 1}}, \quad b, c, k, x > 0$$
+  $$f_X(x;b,c,k) = \frac{{b\,k\,c}}{{x^2}} \frac{{\left( \frac{b}{x} \right)}^{c - 1}}{{\left[ 1 + \left( \frac{b}{x} \right)^c \right]}^{k + 1}}; \quad b, c, k, x > 0$$
   
   <br>
   
-  $$F_X(x;b,c,k) = \frac{1}{{\left[ 1 + \left( \frac{b}{x} \right)^c \right]}^k}, \quad b, c, k, x > 0$$
-  
+  $$F_X(x;b,c,k) = \frac{1}{{\left[ 1 + \left( \frac{b}{x} \right)^c \right]}^k}; \quad b, c, k, x > 0$$
+
 </div>
 
-<div style="text-align: center;"font-size: 1.2em; color:#3664EE;" class="box";>
-<b><u>Burr III distribution</u></b>
-
-$${{f_X}(x;b,c,k) = \frac{{b\,k\,c}}{{{x^2}}}\frac{{{{\left( {\frac{b}{x}} \right)}^{c - 1}}}}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^{k + 1}}}}{\rm{ }}} {b,c,k,x > 0}$$
 <br>
-
-
-$${{F_X}(x;b,c,k) = \frac{1}{{{{\left[ {1 + {{\left( {\frac{b}{x}} \right)}^c}} \right]}^k}}}{\rm{ }}} {b,c,k,x > 0}$$
-</div>
 
 ```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample Burr probability density (A) and cumulative probability (B) functions."}
 par(mfrow = c(1, 2))
[…]
