### 2024-12-02 2762ea3
model-averaging.Rmd latex boxes
