+++ .github/workflows/R-CMD-check.yaml
@@ -57,3 +57,4 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
+
