+++ .github/workflows/R-CMD-check.yaml
@@ -57,4 +57,3 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
-
