+++ dev/pkgdown.yml
@@ -8,4 +8,4 @@ articles:
   distributions: distributions.html
   model-averaging: model-averaging.html
   ssdtools: ssdtools.html
-last_built: 2024-12-02T15:50Z
+last_built: 2024-12-03T15:29Z
