+++ doc/small-sample-bias.R
@@ -1,1 +1,0 @@
-### This is an R script tangled from 'small-sample-bias.pdf.asis'
