### 2024-12-03 fdd30c6
try fixing log-normal math
