+++ vignettes/distributions.Rmd
@@ -258,13 +258,13 @@ A random variable, $X$ is lognormally distributed if the *logarithm* of $X$ is n
 <div style="text-align: center; font-size: 1.2em; color: black; border: 1px solid black; padding: 10px; border-radius: 5px;">
   <b><u>Lognormal Distribution</u></b>
   
-  $${f_X}\left( {x;\mu ,\sigma } \right) = \frac{1}{{\sqrt {2\pi } {\kern 1pt} \sigma x}}\exp \left[ { - \frac{{{{\left( {\ln x - \mu } \right)}^2}}}{{2{\sigma ^2}}}} \right]; \quad x,\sigma  > 0,\; - \infty  < \mu  < \infty $$
+  $${f_X}\left( {x;\mu ,\sigma } \right) = \frac{1}{{\sqrt {2\pi } {\kern 1pt} \sigma x}}\exp \left[ { - \frac{{{{\left( {\ln x - \mu } \right)}^2}}}{{2{\sigma ^2}}}} \right]; \quad x,\sigma  > 0; - \infty  < \mu  < \infty $$
 
 </div>
 
 <br>
 
-The lognormal distribution was selected as the starting distribution given the data are for effect concentrations. 
+The log-normal distribution was selected as the starting distribution given the data are for effect concentrations. 
 The log-normal distribution can be fitted using `ssdtools` supplying the string `lnorm` to the `dists` argument in the `ssd_fit_dists` call.
 
 ```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5,fig.cap="Sample lognormal probability density (A) and cumulative probability (B) functions."}
