### 2024-12-03 38be175
playing with log-normal dist
