### 2024-12-06 ba30c68
add xlimits argument to ssd_plot to set x-axis limits (passed to scale_x_continuous(limits))
