### 2024-12-07 92b12c7
add ssd_label_comma_hc label function to offset hc value (appending \n)
