### 2024-12-07 95de5d2
add ability in plot_coord_scale to append hc value to xbreaks and offset to new line, restrict transformations to log, log10 and identity
