### 2024-12-07 3186e9b
add hc_value param for plot_coord_scale, update documentation on accepted trans values, minor typo in xintercept
