+++ man/params.Rd
@@ -57,6 +57,8 @@ Distributions with an absolute AIC difference greater than delta are excluded fr
 
 \item{hc}{A value between 0 and 1 indicating the proportion hazard concentration (or NULL).}
 
+\item{hc_value}{A number of the hazard concentration value to offset.}
+
 \item{label}{A string of the column in data with the labels.}
 
 \item{left}{A string of the column in data with the concentrations.}
@@ -189,7 +191,7 @@ remove them with a warning.}
 
 \item{tails}{A flag or NULL specifying whether to only include distributions with both tails.}
 
-\item{trans}{A string which transformation to use by default \code{"log10"}.}
+\item{trans}{A string of which transformation to use. Accepted values include \code{"log10"}, \code{"log"}, and \code{"identity"} (\code{"log10"} by default).}
 
 \item{weight}{A string of the numeric column in data with positive weights less than or equal to 1,000 or NULL.}
 
@@ -209,7 +211,7 @@ remove them with a warning.}
 Use NA to refer to the existing minimum or maximum limits.
 }}
 
-\item{xintercept}{The x-value for the intersect}
+\item{xintercept}{The x-value for the intersect.}
 
 \item{xlab}{A string of the x-axis label.}
 
