### 2024-12-07 65ad89e
add two new tests for ssd_plot with hc value
