### 2024-12-07 2fbbd76
add ggtext dependency
