+++ DESCRIPTION
@@ -51,6 +51,7 @@ Imports:
     furrr,
     generics,
     ggplot2,
+    ggtext,
     glue,
     goftest,
     graphics,
