### 2024-12-07 e7fd79e
add text_size and label_size arguments, soft-deprecate size
