+++ R/params.R
@@ -44,6 +44,7 @@
 #' @param hc A value between 0 and 1 indicating the proportion hazard concentration (or NULL).
 #' @param hc_value A number of the hazard concentration value to offset.
 #' @param label A string of the column in data with the labels.
+#' @param label_size A number for the size of the labels.
 #' @param left A string of the column in data with the concentrations.
 #' @param level A number between 0 and 1 of the confidence level of the interval.
 #' @param linecolor A string of the column in pred to use for the line color.
@@ -110,9 +111,10 @@
 #' @param shape2 shape2 parameter.
 #' @param shift_x The value to multiply the label x values by (after adding `add_x`).
 #' @param silent A flag indicating whether fits should fail silently.
-#' @param size A number for the size of the labels.
+#' @param size A number for the size of the labels. Soft-deprecated for `label_size`.
 #' @param suffix Additional text to display after the number on the y-axis.
 #' @param tails A flag or NULL specifying whether to only include distributions with both tails.
+#' @param text_size A number for the text size. 
 #' @param trans A string of which transformation to use. Accepted values include `"log10"`, `"log"`, and `"identity"` (`"log10"` by default).
 #' @param weight A string of the numeric column in data with positive weights less than or equal to 1,000 or NULL.
 #' @param x The object.
