### 2024-12-07 d92ba25
soft-deprecate size for label_size, add option to change text size, format x-axis text as markdown
