### 2024-12-07 3f72201
add snap for label and text size, review/update snapshots with bold hc value
