### 2024-12-07 3503606
forgot to add checks for new xlimits arg
