### 2024-12-07 da95eec
fix xlimits check to allow null
