### 2024-12-07 e7315de
better documentation for ssd_label_comma_hc
