### 2024-12-09 df9fdf6
Added fig alt text, addresses #182
