+++ vignettes/distributions.Rmd
@@ -89,7 +89,7 @@ The probability density function, ${f_X}(x;b,c,k)$ and cumulative distribution f
 
 <br>
 
-```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample Burr probability density (A) and cumulative probability (B) functions."}
+```{r echo=FALSE, fig.align='center', fig.width=9,fig.height=5, fig.cap="Sample Burr probability density (A) and cumulative probability (B) functions.", fig.alt="A two panel plot showing several realisations of the Burr type III probability density function on the left panel and the cumulative density function of distribution on the right panel."}
 par(mfrow = c(1, 2))
 
 f <- function(x, b, c, k) {
@@ -150,7 +150,7 @@ These can be fitted using `ssdtools` by supplying the strings "llogis_llogis" an
 The underlying code for these mixtures has three components: the likelihood function required for TMB; exported R functions to allow the usual methods for a distribution to be called (p, q and r); and a set of supporting R functions (see @fox_methodologies_2022  Appendix D for more details). 
 Both mixtures have five parameters - two parameters for each of the component distributions and a mixing parameter (pmix) that defines the weighting of the two distributions in the ‘mixture.’
 
-```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample lognormal lognormal mixture probability density (A) and cumulative probability (B) functions."}
+```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5, fig.cap="Sample lognormal lognormal mixture probability density (A) and cumulative probability (B) functions.", fig.alt="A two panel plot showing several realisations of the lognormal lognormal mixture probability density function on the left panel and the cumulative density function of distribution on the right panel."}
 par(mfrow = c(1, 2))
 
 f <- function(x, m1, s1, m2, s2, p) {
@@ -216,7 +216,7 @@ The default list of candidate distributions in `ssdtools` is comprised of the fo
 
 The default distributions are plotted below with a mean of 2 and standard deviation of 2 on the (natural) log concentration scale or around 7.4 on the concentration scale.
 
-```{r, message=FALSE, echo=FALSE, fig.width=7,fig.height=5, fig.cap="Currently recommended default distributions."}
+```{r, message=FALSE, echo=FALSE, fig.width=7,fig.height=5, fig.cap="Currently recommended default distributions.", fig.alt="The cumulative distribution function of the six default distributions using a mean and standard deviation of 2 for all. The distributions included are the gammal, lgumbel, llogis, lnorm, lnorm_lnorm and weibull."}
 library(ssdtools)
 library(ggplot2)
 
@@ -266,7 +266,7 @@ A random variable, $X$ is lognormally distributed if the *logarithm* of $X$ is n
 The log-normal distribution was selected as the starting distribution given the data are for effect concentrations. 
 The log-normal distribution can be fitted using `ssdtools` supplying the string `lnorm` to the `dists` argument in the `ssd_fit_dists` call.
 
-```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5,fig.cap="Sample lognormal probability density (A) and cumulative probability (B) functions."}
+```{r echo=FALSE,fig.align='center',fig.width=9,fig.height=5,fig.cap="Sample lognormal probability density (A) and cumulative probability (B) functions.", fig.alt="A two panel plot showing several realisations of the lognormal probability density function on the left panel and the cumulative density function of distribution on the right panel."}
 par(mfrow = c(1, 2))
 
[…]
