+++ scripts/build.R
@@ -31,7 +31,6 @@ lintr::lint_package()
 devtools::test()
 devtools::document()
 
-pkgdown::build_home()
-pkgdown::build_reference()
+pkgdown::build_site()
 
 devtools::check()
