### 2024-12-10 01470fc
add pkgdown::build_reference() for checking.
