+++ README.Rmd
@@ -63,6 +63,7 @@ The website for the development version is at <https://bcgov.github.io/ssdtools/
 ## Introduction
 
 The dependency [`ssddata`](https://github.com/open-AIMS/ssddata) provides example data sets for several chemicals including Boron.
+The [ECOTOX](https://cfpub.epa.gov/ecotox/) Knowledgebase is another source of toxicity concentrations for over 12,000 chemicals.
 
 ```{r, message=FALSE}
 library(ssdtools)
