### 2024-12-10 94effc5
add example for scale_fill_ssd() and expand the description for discrete color-blind scale functions.
