+++ R/ggplot.R
@@ -33,11 +33,14 @@ ssd_pal <- function() {
 }
 
 #' Discrete color-blind scale for SSD Plots
+#' 
+#' The functions were designed for coloring different groups in a plot of SSD data.
 #'
 #' @param ... Arguments passed to [ggplot2::discrete_scale()].
 #' @family ggplot
 #' @export
 #' @examples
+#' # Use the color-blind palette for a SSD plot
 #' ssd_plot(ssddata::ccme_boron, boron_pred, shape = "Group", color = "Group") +
 #'   scale_colour_ssd()
 scale_colour_ssd <- function(...) {
@@ -52,6 +55,12 @@ scale_color_ssd <- function(...) {
 
 #' @describeIn scale_colour_ssd Discrete color-blind scale for SSD Plots
 #' @export
+#' @examples
+#' # Use the color-blind palette for a histogram of concentrations
+#' ggplot2::ggplot(ssddata::ccme_boron, ggplot2::aes(x = Species, y = Conc, fill = Group)) +
+#'   ggplot2::geom_col() +
+#'   scale_fill_ssd() +
+#'   ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust = 0.5, hjust = 1))
 scale_fill_ssd <- function(...) {
   discrete_scale("fill", palette = ssd_pal(), ...)
 }
