### 2024-12-10 5e294b3
update NEWS to clarify that xx refers to family of functions
