### 2024-12-10 9cb7603
Readded FAQ vignette
closes #393
