+++ vignettes/faqs.Rmd
@@ -1,0 +1,46 @@
+---
+title: "Frequently Asked Questions"
+author: "ssdtools Team"
+date: '`r format(Sys.time(), "%Y-%m-%d", tz = "UTC")`'
+bibliography: references.bib
+mathfont: Courier
+latex_engine: MathJax
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Frequently Asked Questions}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 6,
+  fig.height = 4
+)
+```
+
+### How can I plot the model averaged fit and the individual fits?
+
+```{r, alt.text = "A plot showing the model averaged fit and the individual fits for the ccme_boron data."}
+library(ssdtools)
+
+dist <- ssd_fit_dists(ssddata::ccme_boron)
+ssd_plot_cdf(dist, average = NA)
+```
+
+### How can I include mathematical expressions in the xlabel when using `ssd_plot()`?
+
+```{r, message = FALSE, alt.text = "A plot showing the model averaged fit with a mathematical expression in the xlabel with the units."}
+library(ssdtools)
+
+ssd_plot(ssddata::ccme_boron, ssdtools::boron_pred, label = "Species", shape = "Group") +
+  ggplot2::scale_x_continuous(latex2exp::TeX("Boron $(\\mu g$/L)$"), breaks = c(1, 10, 100, 1000, 10000))
+```
+
+<div id="refs"></div>
+
+```{r, results = "asis", echo = FALSE}
+cat(ssd_licensing_md())
+```
