### 2024-12-10 89e9a34
update paper to clarify relationship between multi_est and ci_method
