### 2024-12-10 2c06345
remove argument name without context.
