### 2024-12-10 62f7ca6
update deprecation information
