+++ README.md
@@ -55,8 +55,8 @@ install.packages("ssdtools", repos = c("https://bcgov.r-universe.dev", "https://
 or from [GitHub](https://github.com/bcgov/ssdtools)
 
 ``` r
-# install.packages("pak", repos = sprintf("https://r-lib.github.io/p/pak/stable/%s/%s/%s", .Platform$pkgType, R.Version()$os, R.Version()$arch))
-pak::pak("bcgov/ssdtools")
+# install.packages("remotes")
+remotes::install_github("bcgov/ssdtools")
 ```
 
 #### Website

[…]
