### 2024-12-10 c2c7902
fledge: Bump version to 2.2.0
