### 2024-12-10 2fadc0d
update news file
