+++ vignettes/additional-technical-details.Rmd
@@ -74,7 +74,7 @@ From these, relative bias was calculated as the scaled-difference between the es
 
 Bias, confidence interval width and coverage as a function of sample size across ~1000 simulations of 20 datasets using the multi model averaging method and the weighted sample method for estimating confidence intervals via ssdtools are shown below:
 
-![ An aggregate plot showing how bias and confidence interval width decrease, and covergae increases as the number of data points in the SSD increases. The first three bars are coloured to highlight that the biggest changes are from five to six and then seven data points. ](images/ssdata_sims_collated.png) \
+![ An aggregate plot showing how bias and confidence interval width decrease, and coverage increases as the number of data points in the SSD increases. The first three bars are colored to highlight that the biggest changes are from five to six and then seven data points. ](images/ssdata_sims_collated.png) \
 
 The simulation results showed significant gains in terms of reduced bias from N=5 to N=6, as well as in coverage, which improved substantially between N=5 and N=6. 
 There is also a small additional gain in coverage at N=7, where the median of the simulations reaches the expected 95% but this is only the case for HC1.
@@ -105,7 +105,7 @@ This was examined using the simulation study across the larger combined ssddata 
 Below is a plot of the mean AICc weights as a function of sample size (N) obtained for data simulated using the best fit distribution to 353 datasets. 
 Results are shown separately for the six different source distributions, with the upper plot (A) showing the AICc weight of the source (data generating) distributions, when fit using the default set of six distributions using ssdtools; and the lower plot (B) showing the AICc weight of the lognormal-lognormal mixture distribution for each of the source (data generating) distributions.
 
-![A two panel plot, with both showing how AICc based weights increase as the number of data points in the SSD increases. The top panel shows that when there is a large enough sample size the AICc weights for the true source distribution are high. The bottom panel shows the weights for the lognormal-lognormal mixture distribution and highlights that these remain low except when the lognormal-lognormal misture is the source distribution. ](images/weights_collated.png) \
+![A two panel plot, with both showing how AICc based weights increase as the number of data points in the SSD increases. The top panel shows that when there is a large enough sample size the AICc weights for the true source distribution are high. The bottom panel shows the weights for the lognormal-lognormal mixture distribution and highlights that these remain low except when the lognormal-lognormal mixture is the source distribution. ](images/weights_collated.png) \
 
 We found that the AICc weights for the five unimodal distributions were relatively similar (~0.2) for very low N. 
 This is because for small N it is difficult to discern differences between the distributions in the candidate list.
