+++ vignettes/faqs.Rmd
@@ -23,7 +23,7 @@ knitr::opts_chunk$set(
 
 ### How can I plot the model averaged fit and the individual fits?
 
-```{r, alt.text = "A plot showing the model averaged fit and the individual fits for the ccme_boron data."}
+```{r, fig.alt = "A plot showing the model averaged fit and the individual fits for the ccme_boron data."}
 library(ssdtools)
 
 dist <- ssd_fit_dists(ssddata::ccme_boron)
@@ -32,7 +32,7 @@ ssd_plot_cdf(dist, average = NA)
 
 ### How can I include mathematical expressions in the x-axis label when using `ssd_plot()`?
 
-```{r, message = FALSE, alt.text = "A plot showing the model averaged fit with a mathematical expression in the xlabel with the units."}
+```{r, message = FALSE, fig.alt = "A plot showing the model averaged fit with a mathematical expression in the xlabel with the units."}
 library(ssdtools)
 
 ssd_plot(ssddata::ccme_boron, ssdtools::boron_pred, label = "Species", shape = "Group") +
