### 2024-12-10 7ac705c
simplify text distributions
