+++ paper/paper.md
@@ -54,8 +54,8 @@ The second major release (v2) includes critical updates to ensure that the $\tex
 
 SSDs are a practical tool for the determination of safe threshold concentrations for toxicants in fresh and marine waters, and are implemented in some form for risk assessment and water quality criteria derivation throughout multiple jurisdictions globally [@lepper2005manual; @Warne2018; @bcmecc2019; @USEPA2020].
 
-The selection of a suitable probability model has been identified as one of the most important and difficult choices in the use of SSDs [@newman_applying_2000]. 
-Since the original implementation (v0), `ssdtools` [@thorley2018ssdtools] has used model averaging to allow estimation of $\text{HC}_x$ and $\text{HP}_u$ values using multiple distributions, thereby avoiding the need for selection of a single distribution [@schwarz_improving_2019]. 
+The selection of a suitable distribution has been identified as one of the most important and difficult choices in the use of SSDs [@newman_applying_2000]. 
+Since the original implementation (v0), `ssdtools` [@thorley2018ssdtools] has used model averaging to allow estimation of $\text{HC}_x$ and $\text{HP}_u$ values using multiple distributions [@schwarz_improving_2019]. 
 The method, as applied in the SSD context is described in detail in @fox_recent_2021, and provides a level of flexibility and parsimony that is difficult to achieve with a single distribution.
 
 # Technical details
