### 2024-12-11 #7492 @joethorley
We think we have addressed all your comments/suggestions. We are just doing a final read through of the paper, getting the package ready for a CRAN release (v2.2.0) and ensuring it is compatible with shinyssdtools before provide a formal response to all your comments/suggestions. Thanks for your patience.
