+++ paper/paper.md
@@ -26,7 +26,7 @@ affiliations:
   name: Environmetrics Australia, Australia
 - index: 6
   name: Simon Fraser University, Canada
-date: 10 December 2024
+date: 11 December 2024
 bibliography: paper.bib
 tags:
   - ssdtools
