+++ NEWS.md
@@ -10,6 +10,7 @@
 - For `ssd_plot()` function:
   - Added `text_size` argument.
   - Soft-deprecated `size` argument for `label_size`.
+- Turned off x-axis minor breaks for all plots (for consistency) as HC major break causing multiple minor breaks in `ssd_plot()`.
 
 # ssdtools 2.1.0
 
