+++ R/RcppExports.R
@@ -1,7 +1,7 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-fun <- function(x) {
-    .Call(`_ssdtools_fun`, x)
+fun_ssd <- function(x) {
+    .Call(`_ssdtools_fun_ssd`, x)
 }
 
