+++ R/ssd-plot.R
@@ -39,14 +39,14 @@ plot_coord_scale <- function(data, xlab, ylab, trans, big.mark, suffix, xbreaks 
   list(
     coord_trans(x = trans),
     scale_x_continuous(xlab,
-      breaks = xbreaks,
-      minor_breaks = NULL,
-      labels = ssd_label_fun, 
-      limits = xlimits
+                       breaks = xbreaks,
+                       minor_breaks = NULL,
+                       labels = ssd_label_fun, 
+                       limits = xlimits
     ),
     scale_y_continuous(ylab,
-      labels = label_percent(suffix = suffix), limits = c(0, 1),
-      breaks = seq(0, 1, by = 0.2), expand = c(0, 0)
+                       labels = label_percent(suffix = suffix), limits = c(0, 1),
+                       breaks = seq(0, 1, by = 0.2), expand = c(0, 0)
     )
   )
 }
@@ -70,7 +70,8 @@ ssd_plot <- function(data, pred, left = "Conc", right = left, ...,
                      bounds = c(left = 1, right = 1),
                      big.mark = ",", suffix = "%",
                      trans = "log10", xbreaks = waiver(),
-                     xlimits = NULL, text_size = 11, label_size = 2.5) {
+                     xlimits = NULL, text_size = 11, label_size = 2.5,
+                     theme_classic = FALSE) {
   
   if (lifecycle::is_present(size)) {
     lifecycle::deprecate_soft("2.1.0", "ssd_plot(size)", "ssd_plot(label_size)", id = "size")
@@ -85,20 +86,20 @@ ssd_plot <- function(data, pred, left = "Conc", right = left, ...,
   chk_null_or(linetype, vld = vld_string)
   chk_null_or(linecolor, vld = vld_string)
   check_names(data, c(unique(c(left, right)), label, shape))
-
+  
   check_names(pred, c("proportion", "est", "lcl", "ucl", unique(c(linetype, linecolor))))
   chk_numeric(pred$proportion)
   chk_range(pred$proportion)
   check_data(pred, values = list(est = 1, lcl = c(1, NA), ucl = c(1, NA)))
-
+  
   chk_number(shift_x)
   chk_range(shift_x, c(1, 1000))
   chk_number(add_x)
   chk_range(add_x, c(-1000, 1000))
-
+  
   chk_flag(ci)
   chk_flag(ribbon)
-
+  
   if (!is.null(hc)) {
     chk_vector(hc)
     chk_gt(length(hc))
@@ -111,19 +112,20 @@ ssd_plot <- function(data, pred, left = "Conc", right = left, ...,
   chk_number(text_size)
   chk_null_or(xlimits, vld = vld_numeric)
   chk_null_or(xlimits, vld = vld_length, length = 2L)
-
+  chk_flag(theme_classic)
+  
   data <- process_data(data, left, right, weight = NULL)
   data <- bound_data(data, bounds)
   data$y <- ssd_ecd_data(data, "left", "right", bounds = bounds)
-
+  
   label <- if (!is.null(label)) sym(label) else label
   shape <- if (!is.null(shape)) sym(shape) else shape
   color <- if (!is.null(color)) sym(color) else color
   linetype <- if (!is.null(linetype)) sym(linetype) else linetype
   linecolor <- if (!is.null(linecolor)) sym(linecolor) else linecolor
-
+  
   gp <- ggplot(data)
-
+  
   if (ci) {
     if (ribbon) {
       gp <- gp + geom_xribbon(data = pred, aes(xmin = !!sym("lcl"), xmax = !!sym("ucl"), y = !!sym("proportion")), alpha = 0.2)
[…]
