### 2024-12-16 #7492 @fabian-s
thank you @joethorley for this structured and prompt response!

@flor14 @nanhung
please cross off the open points on your checklists that have been successfully addressed by these changes. For the ones that remain open -- if any -- please provide some additional info what else would need to be done in your opinion.
