### 2024-12-19 1f0b3ca
added `ssd_dists_shiny()`
