+++ man/ssd_dists_shiny.Rd
@@ -1,24 +1,25 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/dists.R
-\name{ssd_dists_all}
-\alias{ssd_dists_all}
-\title{All Species Sensitivity Distributions}
+\name{ssd_dists_shiny}
+\alias{ssd_dists_shiny}
+\title{All Shiny Species Sensitivity Distributions}
 \usage{
-ssd_dists_all()
+ssd_dists_shiny()
 }
 \value{
 A unique, sorted character vector of the distributions.
 }
 \description{
-Gets a character vector of the names of all the available distributions.
+Gets a character vector of the names of all the available distributions
+in the shinyssdtools.
 }
 \examples{
-ssd_dists_all()
+ssd_dists_shiny()
 }
 \seealso{
 Other dists: 
 \code{\link{dist_data}},
 \code{\link{ssd_dists}()},
-\code{\link{ssd_dists_shiny}()}
+\code{\link{ssd_dists_all}()}
 }
 \concept{dists}
