+++ NAMESPACE
@@ -75,6 +75,7 @@ export(ssd_data)
 export(ssd_dists)
 export(ssd_dists_all)
 export(ssd_dists_bcanz)
+export(ssd_dists_shiny)
 export(ssd_eburrIII3)
 export(ssd_ecd)
 export(ssd_ecd_data)
