### 2024-12-27 bcgov/ssdtools#387 @flor14
@joethorley  I'm reviewing the documentation, and it seems there are still issues with some images and formulas. I'm checking the webpage linked to the repository: https://bcgov.github.io/ssdtools/
Is this the correct link?

![Image](https://github.com/user-attachments/assets/4f8dc0d1-fcc0-4a95-b644-1a29d4b15494)
![Image](https://github.com/user-attachments/assets/a2b8fd70-acbd-43aa-9cc6-55667eff5d9c)
