### 2024-12-27 #7492 @flor14
@joethorley The vignettes are very clear and complete. There is just a small detail left in this log-normal distribution (I am checking the `dev/` version of the docs so I assume this is not corrected, could be?):
