### 2024-12-28 #7492 @flor14
If this issue I mentioned about the log-normal is resolved, the review is complete for me. I’ve already checked the other points.
