### 2025-01-02 #7492 @fabian-s
@nanhung
please cross off the open points on your checklists that have been successfully addressed by the changes @joethorley listed above.
For the ones that remain open -- if any -- please provide some additional info what else would need to be done in your opinion.
