### 2025-01-02 #7492 @joethorley
Thanks @flor14 and @nanhung.

@fabian-s - I am currently in the process of updating on the main branch to update the website and submitting the lastest version to CRAN. This may take a week to 10 days as I need approval to update the main branch on the bcgov site but the approver is on vacation. Once this is complete I'll let you know and the paper should be ready for finalization. Thanks everyone!
