+++ pkgdown.yml
@@ -6,6 +6,10 @@ articles:
   articles/confidence-intervals: confidence-intervals.html
   customising-plots: customising-plots.html
   distributions: distributions.html
+  faqs: faqs.html
   model-averaging: model-averaging.html
   ssdtools: ssdtools.html
-last_built: 2024-10-22T15:26Z
+last_built: 2025-01-06T22:30Z
+urls:
+  reference: https://bcgov.github.io/ssdtools/reference
+  article: https://bcgov.github.io/ssdtools/articles
