+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   faqs: faqs.html
   model-averaging: model-averaging.html
   ssdtools: ssdtools.html
-last_built: 2025-01-06T22:30Z
+last_built: 2025-01-07T11:59Z
 urls:
   reference: https://bcgov.github.io/ssdtools/reference
   article: https://bcgov.github.io/ssdtools/articles
