### 2025-01-07 #7492 @fabian-s
@joethorley I have no further comments/corrections re. the text of the paper, please finalize the submission once the CRAN submission is done, see the "Additional Author Tasks " above.
