+++ vignettes/articles/additional-technical-details.Rmd
@@ -2,7 +2,7 @@
 title: "Additional Technical Details"
 author: "ssdtools Team"
 date: '`r format(Sys.time(), "%Y-%m-%d", tz = "UTC")`'
-bibliography: references.bib
+bibliography: ../references.bib
 mathfont: Courier
 latex_engine: MathJax
 output: rmarkdown::html_vignette
