+++ _pkgdown.yml
@@ -131,5 +131,6 @@ articles:
   - articles/distributions
   - articles/confidence-intervals
   - articles/customising-plots
+  - articles/small-sample-bias
   - articles/additional-technical-details
   - faqs
