### 2025-01-14 db0df0a
Updated _pkgdown.yml to new Poisson template
