+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   articles/model-averaging: model-averaging.html
   articles/small-sample-bias: small-sample-bias.html
   ssdtools: ssdtools.html
-last_built: 2025-01-14T17:37Z
+last_built: 2025-01-14T20:19Z
 urls:
   reference: https://bcgov.github.io/ssdtools/reference
   article: https://bcgov.github.io/ssdtools/articles
