### 2025-01-14 #7492 @joethorley
The release version number which is now on CRAN is v2.2.0 https://cran.r-project.org/web/packages/ssdtools/index.html

The DOI for the Zenodo release is https://doi.org/10.5281/zenodo.14609128

I've done all the other checks.

@fabian-s - Let us know if you need anything else and thanks for your patience.
