### 2025-01-14 c131a3d
Updated DESCRIPTION to new Poisson template
