### 2025-01-15 #7492 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
