### 2025-01-15 3b49e86
Undo _pkgdown.yml changes
