+++ DESCRIPTION
@@ -110,4 +110,3 @@ Language: en-US
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
-Config/Needs/website: poissonconsulting/poissontemplate
