### 2025-01-28 #7492 @Kevin-Mattheus-Moerman
> If this issue I mentioned about the log-normal is resolved, the review is complete for me. I’ve already checked the other points.

As per this comment by @flor14 I will assume the remaining box on the `Functionality documentation` can now be ticked.
